requirements.txt
Django==4.2.14
pillow==10.3.0
psycopg2-binary==2.9.9
django-bootstrap5==24.3
djangorestframework==3.14.0
drf-yasg==1.21.7